fix(circleci): big refactor and cleanup
