+++ plugins/circleci/src/components/BuildsPage/BuildsPage.tsx
@@ -13,4 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './Layout';
+import React from 'react';
+import { Builds } from './lib/Builds';
+import { Grid } from '@material-ui/core';
+
+export const BuildsPage = () => (
+  <Grid container spacing={3} direction="column">
+    <Grid item>
+      <Builds />
+    </Grid>
+  </Grid>
+);
