+++ plugins/circleci/src/state/index.ts
@@ -13,7 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './AppState';
-export * from './useSettings';
 export * from './useBuilds';
 export * from './useBuildWithSteps';
